requirements.txt
apache-airflow==2.8.1
apache-airflow-providers-common-io==1.2.0
apache-airflow-providers-common-sql==1.10.1
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-http==4.9.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-sqlite==3.7.0
beautifulsoup4==4.12.3
boto3==1.34.106
botocore==1.34.106
mysql-connector-python==8.1.0
mysql-replication==0.22
matplotlib==3.7.2
nltk==3.8.1
numpy==1.25.2
pandas==2.0.3
pymysql==0.10.1
pymongo==4.6.1
pyspark==3.5.0
requests==2.31.0
requests-toolbelt==1.0.0
s3fs==2023.10.0
SQLAlchemy==1.4.51
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
urllib3==2.1.0